alterando update e delete no controller animal

+++ back/src/main/java/br/edu/ufape/hvu/controller/AnimalController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.jwt.Jwt;
@@ -61,9 +62,14 @@ public class AnimalController {
 	@PatchMapping("animal/{id}")
 	public AnimalResponse updateAnimal(@PathVariable Long id, @Valid @RequestBody AnimalRequest obj) {
 		try {
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			Jwt principal = (Jwt) authentication.getPrincipal();
 			//Animal o = obj.convertToEntity();
 			Animal oldObject = facade.findAnimalById(id);
-
+			Tutor tutor = facade.findTutorByanimalId(oldObject.getId());
+			if(!principal.getSubject().equals(tutor.getUserId())) {
+				throw new AccessDeniedException("This is not your animal");
+			}
 			TypeMap<AnimalRequest, Animal> typeMapper = modelMapper
 													.typeMap(AnimalRequest.class, Animal.class)
 													.addMappings(mapper -> mapper.skip(Animal::setId));			
@@ -80,6 +86,13 @@ public class AnimalController {
 	@DeleteMapping("animal/{id}")
 	public String deleteAnimal(@PathVariable Long id) {
 		try {
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			Jwt principal = (Jwt) authentication.getPrincipal();
+			Animal oldObject = facade.findAnimalById(id);
+			Tutor tutor = facade.findTutorByanimalId(oldObject.getId());
+			if(!principal.getSubject().equals(tutor.getUserId())) {
+				throw new AccessDeniedException("This is not your animal");
+			}
 			facade.deleteAnimal(id);
 			return "";
 		} catch (RuntimeException ex) {
